requirements.txt
Django==5.1.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
